python/pyproject.toml
[project]
name = "otf-api"
version = "0.15.4"
description = "Python OrangeTheory Fitness API Client"
authors = [{ name = "Jessica Smith", email = "j.smith.git1@gmail.com" }]
requires-python = ">=3.11"
readme = "README.md"
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Internet :: WWW/HTTP",
    "Operating System :: OS Independent",
]
dependencies = [
    "attrs>=24.3.0,<25",
    "httpx>=0.27.2",
    "humanize>=4.9.0,<5",
    "inflection==0.5.*",
    "pint==0.24.*",
    "pycognito==2024.5.1",
    "pydantic>=2.7.3",
    "yarl>=1.18.3,<2",
    "tenacity>=9.0.0,<10",
    "cachetools>=5.5.0",
    "pendulum>=3.1.0",
    "diskcache>=5.6.3",
    "platformdirs>=4.3.6",
    "packaging>=24.2",
    "coloredlogs>=15.0.1",
    "pyyaml>=6.0.0",
]

[project.urls]
Documentation = "https://otf-api.readthedocs.io/en/stable/"

[dependency-groups]
dev = [
    "build==1.2.1",
    "bump-my-version>=0.23.0,<0.24",
    "mypy-boto3-cognito-identity>=1.38.0",
    "mypy-boto3-cognito-idp>=1.35.93,<2",
    "pre-commit==3.7.1",
    "pytest==8.2.2",
    "pytest-cov==5.0.0",
    "tox==4.15.1",
    "twine==5.1.1",
]
docs = [
    "autodoc-pydantic>=2.2.0",
    "furo>=2024.8.6",
    "sphinx>=8.3.0",
    "sphinx-autobuild==2024.10.3",
    "sphinx-autodoc-typehints==3.2.0",
    "sphinx-autodoc2==0.5.0",
    "sphinx-basic-ng==1.0.0b2",
    "sphinxcontrib-applehelp==2.0.0",
    "sphinxcontrib-devhelp==2.0.0",
    "sphinxcontrib-htmlhelp==2.1.0",
    "sphinxcontrib-jsmath==1.0.1",
    "sphinxcontrib-qthelp==2.0.0",
    "sphinxcontrib-serializinghtml==2.0.0",
]

[tool.uv]
default-groups = ["dev"]


[tool.pytest.ini_options]
pythonpath = ["src"]

[build-system]
requires = ["setuptools >= 80.0"]
build-backend = "setuptools.build_meta"
